+++ e2e/dbaas_mysql/resource_mysql.go
@@ -1,4 +1,4 @@
-package dbaasmysql
+package dbaas_mysql
 
 import (
 	"context"
