+++ e2e/dbaas_postgress/resource_postgress.go
@@ -1,4 +1,4 @@
-package postgres
+package dbaas_postgress
 
 import (
 	"context"
