+++ e2e/provider.go
@@ -3,8 +3,8 @@ package e2e
 import (
 	"github.com/e2eterraformprovider/terraform-provider-e2e/client"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/blockstorage"
-	dbaasmysql "github.com/e2eterraformprovider/terraform-provider-e2e/e2e/dbaas_mysql"
-	postgres "github.com/e2eterraformprovider/terraform-provider-e2e/e2e/dbaas_postgress"
+	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/dbaas_mysql"
+	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/dbaas_postgress"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/image"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/kubernetes"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/loadbalancer"
@@ -53,8 +53,8 @@ func Provider() *schema.Provider {
 			"e2e_objectstore":   objectstore.ResourceObjectStore(),
 			"e2e_ssh_key":       ssh_key.ResourceSshKey(),
 			"e2e_kubernetes":    kubernetes.ResourceKubernetesService(),
-			"e2e_postgresDBaaS": postgres.ResourcePostgresDBaaS(),
-			"e2e_dbaas_mysql":   dbaasmysql.ResourceMySql(),
+			"e2e_postgresDBaaS": dbaas_postgress.ResourcePostgresDBaaS(),
+			"e2e_dbaas_mysql":   dbaas_mysql.ResourceMySql(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"e2e_node":            node.DataSourceNode(),
@@ -69,8 +69,8 @@ func Provider() *schema.Provider {
 			"e2e_sfss":            sfs.DataSourceSfs(),
 			"e2e_objectstores":    objectstore.DataSourceObjectStores(),
 			"e2e_kubernetes":      kubernetes.DataSourceKubernetesService(),
-			"e2e_postgresDBaaS":   postgres.DataSourcePostgresDBaaS(),
-			"e2e_dbaas_mysql":     dbaasmysql.DataSourceMySQLDBaaS(),
+			"e2e_postgresDBaaS":   dbaas_postgress.DataSourcePostgresDBaaS(),
+			"e2e_dbaas_mysql":     dbaas_mysql.DataSourceMySQLDBaaS(),
 		},
 		ConfigureFunc: providerConfigure,
 	}
